refactor : 회원관리 리팩토링
스웨거 정리

+++ E101_PJT_1/backend/src/main/java/com/ssafy/api/controller/UserController.java
@@ -1,5 +1,6 @@
 package com.ssafy.api.controller;
 
+import org.hibernate.PropertyValueException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -7,7 +8,6 @@ import org.springframework.security.core.Authentication;
 import org.springframework.web.bind.annotation.*;
 
 import com.ssafy.api.request.UserRegisterPostReq;
-import com.ssafy.api.response.UserRes;
 import com.ssafy.api.service.UserService;
 import com.ssafy.common.auth.SsafyUserDetails;
 import com.ssafy.db.entity.User;
@@ -20,6 +20,7 @@ import io.swagger.annotations.ApiResponses;
 import springfox.documentation.annotations.ApiIgnore;
 
 import java.util.List;
+import java.util.Optional;
 
 /**
  * 유저 관련 API 요청 처리를 위한 컨트롤러 정의.
@@ -33,66 +34,81 @@ public class UserController {
 	@Autowired
 	UserService userService;
 
-	@GetMapping
-	public List<User> getAllUsers() {
-		return userService.getAllUsers();
+	@ExceptionHandler(PropertyValueException.class)
+	@ResponseStatus(HttpStatus.NOT_FOUND)
+	public String handle404(PropertyValueException ex) {
+		return "잘못된 body 입니다";
 	}
+
 	
-	@PostMapping(produces ="text/plain; charset=utf-8")
+	@PostMapping
 	@ApiOperation(value = "회원 가입", notes = "<strong>아이디와 패스워드</strong>를 통해 회원가입 한다.") 
     @ApiResponses({
         @ApiResponse(code = 200, message = "성공"),
-        @ApiResponse(code = 401, message = "인증 실패"),
-        @ApiResponse(code = 404, message = "사용자 없음"),
+        @ApiResponse(code = 400, message = "잘못된 요청"),
+		@ApiResponse(code = 404, message = "바디 정보 오류"),
         @ApiResponse(code = 500, message = "서버 오류")
     })
-	public void register(
+	public ResponseEntity<?> register(
 			@RequestBody @ApiParam(value="회원가입 정보", required = true) UserRegisterPostReq registerInfo) {
-		
-		//임의로 리턴된 User 인스턴스. 현재 코드는 회원 가입 성공 여부만 판단하기 때문에 굳이 Insert 된 유저 정보를 응답하지 않음.
-		User user = userService.createUser(registerInfo);
+		String userId = registerInfo.getUserId();
+		String nickname = registerInfo.getNickname();
+		if (userService.checkIdDuplicated(userId) && userService.checkNicknameDuplicated(nickname)) {
+			User user = userService.createUser(registerInfo);
+			return new ResponseEntity<>(userId + "의 회원가입이 완료되었습니다", HttpStatus.valueOf(200));
+		}
+		return new ResponseEntity<>("잘못된 요청입니다", HttpStatus.valueOf(400));
 	}
 	
 	@GetMapping("my-info")
 	@ApiOperation(value = "회원 본인 정보 조회", notes = "로그인한 회원 본인의 정보를 응답한다.") 
     @ApiResponses({
         @ApiResponse(code = 200, message = "성공"),
-        @ApiResponse(code = 401, message = "인증 실패"),
-        @ApiResponse(code = 404, message = "사용자 없음"),
+        @ApiResponse(code = 400, message = "사용자 없음"),
+		@ApiResponse(code = 401, message = "인증 실패"),
+		@ApiResponse(code = 403, message = "토큰 없음"),
         @ApiResponse(code = 500, message = "서버 오류")
     })
-	public ResponseEntity<User> getUserInfo(@ApiIgnore Authentication authentication) {
-		/**
-		 * 요청 헤더 액세스 토큰이 포함된 경우에만 실행되는 인증 처리이후, 리턴되는 인증 정보 객체(authentication) 통해서 요청한 유저 식별.
-		 * 액세스 토큰이 없이 요청하는 경우, 403 에러({"error": "Forbidden", "message": "Access Denied"}) 발생.
-		 */
+	public ResponseEntity<?> getUserInfo(@ApiIgnore Authentication authentication) {
+
+		if (authentication == null) {
+			return new ResponseEntity<>("토큰이 없습니다", HttpStatus.valueOf(403));
+		}
 		SsafyUserDetails userDetails = (SsafyUserDetails)authentication.getDetails();
 		String userId = userDetails.getUsername();
 		User user = userService.getUserByUserId(userId);
-		
-		return new ResponseEntity<User>(user, HttpStatus.OK);
+		if (user != null) {
+			return new ResponseEntity<User>(user, HttpStatus.valueOf(200));
+		}
+		return new ResponseEntity<>("잘못된 요청입니다", HttpStatus.valueOf(400));
 	}
 
 	@PutMapping(value="my-info")
 	@ApiOperation(value = "회원 본인 정보 수정", notes = "로그인한 회원 본인의 정보를 수정한다.")
 	@ApiResponses({
 			@ApiResponse(code = 200, message = "성공"),
+			@ApiResponse(code = 400, message = "잘못된 요청"),
 			@ApiResponse(code = 401, message = "인증 실패"),
-			@ApiResponse(code = 404, message = "사용자 없음"),
+			@ApiResponse(code = 403, message = "토큰 없음"),
+			@ApiResponse(code = 404, message = "바디 정보 오류"),
+			@ApiResponse(code = 405, message = "무결성 오류"),
 			@ApiResponse(code = 500, message = "서버 오류")
 	})
-	public ResponseEntity<User> updateUserInfo(@ApiIgnore Authentication authentication,
+	public ResponseEntity<?> updateUserInfo(@ApiIgnore Authentication authentication,
 												  @RequestBody @ApiParam(value="회원가입 정보", required = true) UserRegisterPostReq updateInfo) {
-		/**
-		 * 요청 헤더 액세스 토큰이 포함된 경우에만 실행되는 인증 처리이후, 리턴되는 인증 정보 객체(authentication) 통해서 요청한 유저 식별.
-		 * 액세스 토큰이 없이 요청하는 경우, 403 에러({"error": "Forbidden", "message": "Access Denied"}) 발생.
-		 */
+
+		if (authentication == null) {
+			return new ResponseEntity<>("토큰이 없습니다", HttpStatus.valueOf(403));
+		}
 		SsafyUserDetails userDetails = (SsafyUserDetails)authentication.getDetails();
 		String userId = userDetails.getUsername();
-		User user = userService.updateUserInfo(userId, updateInfo);
-		System.out.println(user.toString());
-
-		return new ResponseEntity<>(user, HttpStatus.OK);
+		if (userId.equals(updateInfo.getUserId())) {
+			User user = userService.updateUserInfo(userId, updateInfo);
+			if (user == null)
+				return new ResponseEntity<>("무결성 오류입니다", HttpStatus.valueOf(405));
+			return new ResponseEntity<>(user, HttpStatus.valueOf(200));
+		}
+		return new ResponseEntity<>("잘못된 요청입니다", HttpStatus.valueOf(400));
 	}
 
 	@DeleteMapping(value="my-info")
@@ -100,27 +116,39 @@ public class UserController {
 	@ApiResponses({
 			@ApiResponse(code = 200, message = "성공"),
 			@ApiResponse(code = 401, message = "인증 실패"),
+			@ApiResponse(code = 403, message = "토큰 없음"),
 			@ApiResponse(code = 404, message = "사용자 없음"),
 			@ApiResponse(code = 500, message = "서버 오류")
 	})
-	public void deleteUser(@ApiIgnore Authentication authentication) {
-		/**
-		 * 요청 헤더 액세스 토큰이 포함된 경우에만 실행되는 인증 처리이후, 리턴되는 인증 정보 객체(authentication) 통해서 요청한 유저 식별.
-		 * 액세스 토큰이 없이 요청하는 경우, 403 에러({"error": "Forbidden", "message": "Access Denied"}) 발생.
-		 */
+	public ResponseEntity<?> deleteUser(@ApiIgnore Authentication authentication) {
+		if (authentication == null) {
+			return new ResponseEntity<>("토큰이 없습니다", HttpStatus.valueOf(403));
+		}
 		SsafyUserDetails userDetails = (SsafyUserDetails)authentication.getDetails();
 		String userId = userDetails.getUsername();
 		User user = userService.getUserByUserId(userId);
+		if (user == null) {
+			return new ResponseEntity<>(userId + "의 회원 정보가 없습니다", HttpStatus.valueOf(404));
+		}
 		userService.deleteUser(user);
+		return new ResponseEntity<>(userId + "의 회원 정보가 삭제되었습니다", HttpStatus.valueOf(200));
 	}
 
 	@GetMapping("check-id/{userId}")
+	@ApiResponses({
+			@ApiResponse(code = 200, message = "성공"),
+			@ApiResponse(code = 500, message = "서버 오류")
+	})
 	@ApiOperation(value = "아이디 중복 체크", notes = "중복이면 false, 유효하면 true")
 	public ResponseEntity<Boolean> checkId(@PathVariable("userId") String userId) {
 		return new ResponseEntity<Boolean>(userService.checkIdDuplicated(userId), HttpStatus.OK);
 	}
 
 	@GetMapping("check-nickname/{nickname}")
+	@ApiResponses({
+			@ApiResponse(code = 200, message = "성공"),
+			@ApiResponse(code = 500, message = "서버 오류")
+	})
 	@ApiOperation(value = "닉네임 중복 체크", notes = "중복이면 false, 유효하면 true")
 	public ResponseEntity<Boolean> checkNickname(@PathVariable("nickname") String nickname) {
 		return new ResponseEntity<Boolean>(userService.checkNicknameDuplicated(nickname), HttpStatus.OK);
